+++ bmc-core/src/main/java/com/oracle/bmc/core/ComputeManagement.java
@@ -47,6 +47,40 @@ public interface ComputeManagement extends AutoCloseable {
     AttachLoadBalancerResponse attachLoadBalancer(AttachLoadBalancerRequest request);
 
     /**
+     * Moves an instance configuration into a different compartment within the same tenancy.
+     * For information about moving resources between compartments, see
+     * [Moving Resources to a Different Compartment](https://docs.cloud.oracle.com/iaas/Content/Identity/Tasks/managingcompartments.htm#moveRes).
+     * <p>
+     **Important:** Most of the properties for an existing instance configuration, including the compartment,
+     * cannot be modified after you create the instance configuration. Although you can move an instance configuration
+     * to a different compartment, you will not be able to use the instance configuration to manage instance pools
+     * in the new compartment. If you want to update an instance configuration to point to a different compartment,
+     * you should instead create a new instance configuration in the target compartment using
+     * [CreateInstanceConfiguration](https://docs.cloud.oracle.com/iaas/api/#/en/iaas/20160918/InstanceConfiguration/CreateInstanceConfiguration).
+     *
+     * @param request The request object containing the details to send
+     * @return A response object containing details about the completed operation
+     * @throws BmcException when an error occurs.
+     */
+    ChangeInstanceConfigurationCompartmentResponse changeInstanceConfigurationCompartment(
+            ChangeInstanceConfigurationCompartmentRequest request);
+
+    /**
+     * Moves an instance pool into a different compartment within the same tenancy. For
+     * information about moving resources between compartments, see
+     * [Moving Resources to a Different Compartment](https://docs.cloud.oracle.com/iaas/Content/Identity/Tasks/managingcompartments.htm#moveRes).
+     * <p>
+     * When you move an instance pool to a different compartment, associated resources such as the instances in
+     * the pool, boot volumes, VNICs, and autoscaling configurations are not moved.
+     *
+     * @param request The request object containing the details to send
+     * @return A response object containing details about the completed operation
+     * @throws BmcException when an error occurs.
+     */
+    ChangeInstancePoolCompartmentResponse changeInstancePoolCompartment(
+            ChangeInstancePoolCompartmentRequest request);
+
+    /**
      * Creates an instance configuration
      *
      * @param request The request object containing the details to send