+++ bmc-core/src/main/java/com/oracle/bmc/core/VirtualNetworkAsync.java
@@ -106,6 +106,25 @@ public interface VirtualNetworkAsync extends AutoCloseable {
                             handler);
 
     /**
+     * Moves a NAT gateway into a different compartment within the same tenancy. For information
+     * about moving resources between compartments, see
+     * [Moving Resources to a Different Compartment](https://docs.cloud.oracle.com/iaas/Content/Identity/Tasks/managingcompartments.htm#moveRes).
+     *
+     *
+     * @param request The request object containing the details to send
+     * @param handler The request handler to invoke upon completion, may be null.
+     * @return A Future that can be used to get the response if no AsyncHandler was
+     *         provided. Note, if you provide an AsyncHandler and use the Future, some
+     *         types of responses (like java.io.InputStream) may not be able to be read in
+     *         both places as the underlying stream may only be consumed once.
+     */
+    java.util.concurrent.Future<ChangeNatGatewayCompartmentResponse> changeNatGatewayCompartment(
+            ChangeNatGatewayCompartmentRequest request,
+            com.oracle.bmc.responses.AsyncHandler<
+                            ChangeNatGatewayCompartmentRequest, ChangeNatGatewayCompartmentResponse>
+                    handler);
+
+    /**
      * Moves a service gateway into a different compartment within the same tenancy. For information
      * about moving resources between compartments, see
      * [Moving Resources to a Different Compartment](https://docs.cloud.oracle.com/iaas/Content/Identity/Tasks/managingcompartments.htm#moveRes).
@@ -125,25 +144,6 @@ public interface VirtualNetworkAsync extends AutoCloseable {
                                     ChangeServiceGatewayCompartmentRequest,
                                     ChangeServiceGatewayCompartmentResponse>
                             handler);
-
-    /**
-     * Moves a NAT gateway into a different compartment within the same tenancy. For information
-     * about moving resources between compartments, see
-     * [Moving Resources to a Different Compartment](https://docs.cloud.oracle.com/iaas/Content/Identity/Tasks/managingcompartments.htm#moveRes).
-     *
-     *
-     * @param request The request object containing the details to send
-     * @param handler The request handler to invoke upon completion, may be null.
-     * @return A Future that can be used to get the response if no AsyncHandler was
-     *         provided. Note, if you provide an AsyncHandler and use the Future, some
-     *         types of responses (like java.io.InputStream) may not be able to be read in
-     *         both places as the underlying stream may only be consumed once.
-     */
-    java.util.concurrent.Future<ChangeNatGatewayCompartmentResponse> changeNatGatewayCompartment(
-            ChangeNatGatewayCompartmentRequest request,
-            com.oracle.bmc.responses.AsyncHandler<
-                            ChangeNatGatewayCompartmentRequest, ChangeNatGatewayCompartmentResponse>
-                    handler);
 
     /**
      * Connects this local peering gateway (LPG) to another one in the same region.