+++ bmc-core/src/main/java/com/oracle/bmc/core/model/InstanceConfigurationInstanceSourceViaImageDetails.java
@@ -86,7 +86,7 @@ public class InstanceConfigurationInstanceSourceViaImageDetails
     }
 
     /**
-     * The size of the boot volume in GBs. Minimum value is 50 GB and maximum value is 16384 GB (16TB).
+     * The size of the boot volume in GBs. The minimum value is 50 GB and the maximum value is 16384 GB (16TB).
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("bootVolumeSizeInGBs")
     Long bootVolumeSizeInGBs;