+++ bmc-streaming/src/main/java/com/oracle/bmc/streaming/StreamAdminClient.java
@@ -18,8 +18,7 @@ public class StreamAdminClient implements StreamAdmin {
             com.oracle.bmc.Services.serviceBuilder()
                     .serviceName("STREAMADMIN")
                     .serviceEndpointPrefix("streams")
-                    .serviceEndpointTemplate(
-                            "https://streams.{region}.streaming.oci.{secondLevelDomain}")
+                    .serviceEndpointTemplate("https://streaming.{region}.oci.{secondLevelDomain}")
                     .build();
     // attempt twice if it's instance principals, immediately failures will try to refresh the token
     private static final int MAX_IMMEDIATE_RETRIES_IF_USING_INSTANCE_PRINCIPALS = 2;